split context component and useContext hook :(
